+++ src/Widgets/Controls/Button.cpp
@@ -13,10 +13,10 @@ Button::Button(Graphic* graphic, const QString& text, bool auto_touch_target_com
 {
 	init();
 	
-	a_corner_radius_tl.set_value(7);
-	a_corner_radius_tr.set_value(7);
-	a_corner_radius_bl.set_value(7);
-	a_corner_radius_br.set_value(7);
+	a_corner_radius_tl.set_value(7.0);
+	a_corner_radius_tr.set_value(7.0);
+	a_corner_radius_bl.set_value(7.0);
+	a_corner_radius_br.set_value(7.0);
 	a_hover_fill.set_disabled(false);
 }
 
@@ -34,10 +34,10 @@ Button::Button(const QString& text, bool auto_touch_target_compliance, QWidget* 
 {
 	init();
 
-	a_corner_radius_tl.set_value(7);
-	a_corner_radius_tr.set_value(7);
-	a_corner_radius_bl.set_value(7);
-	a_corner_radius_br.set_value(7);
+	a_corner_radius_tl.set_value(7.0);
+	a_corner_radius_tr.set_value(7.0);
+	a_corner_radius_bl.set_value(7.0);
+	a_corner_radius_br.set_value(7.0);
 	a_hover_fill.set_disabled(false);
 }
 
@@ -128,7 +128,7 @@ void Button::init_child_themeable_reference_list()
 
 void Button::resize()
 {
-	int border_thickness = a_border_thickness.as<int>();
+	int border_thickness = a_border_thickness.as<double>();
 	int content_height = 0;
 	int graphic_width = 0;
 	int layout_spacing = 0;
