+++ src/Widgets/Dialogs/CreateNewThemeDialog.cpp
@@ -79,15 +79,15 @@ CreateNewThemeDialog::CreateNewThemeDialog(QWidget* parent) : QDialog(parent)
 	m_start_theme_combobox->set_font_size(15);
 
 	connect(&a_border_thickness, &Attribute::value_changed, [this] {
-		int margin = a_border_thickness.as<int>();
+		int margin = a_border_thickness.as<double>();
 
 		m_main_layout->setContentsMargins(margin, margin, margin, margin);
 
-		//a_corner_radius_tl.set_value(inner_radius(a_corner_radius_tl.as<int>(), a_border_thickness.as<int>()));
-		//a_corner_radius_tr.set_value(inner_radius(a_corner_radius_tr.as<int>(), a_border_thickness.as<int>()));
+		//a_corner_radius_tl.set_value(inner_radius(a_corner_radius_tl.as<double>(), a_border_thickness.as<double>()));
+		//a_corner_radius_tr.set_value(inner_radius(a_corner_radius_tr.as<double>(), a_border_thickness.as<double>()));
 
-		m_titlebar->a_corner_radius_tl.set_value(inner_radius(a_corner_radius_tl.as<int>(), a_border_thickness.as<int>()));
-		m_titlebar->a_corner_radius_tr.set_value(inner_radius(a_corner_radius_tr.as<int>(), a_border_thickness.as<int>()));
+		m_titlebar->a_corner_radius_tl.set_value(inner_radius(a_corner_radius_tl.as<double>(), a_border_thickness.as<double>()));
+		m_titlebar->a_corner_radius_tr.set_value(inner_radius(a_corner_radius_tr.as<double>(), a_border_thickness.as<double>()));
 		});
 
 	setup_layout();
@@ -166,20 +166,20 @@ void CreateNewThemeDialog::init_attributes()
 	m_attribute_layout.append(&a_corner_color);
 	m_attribute_layout.append(&a_outline_color);
 
-	m_titlebar->a_corner_radius_tl.set_value(inner_radius(a_corner_radius_tl.as<int>(), a_border_thickness.as<int>()));
-	m_titlebar->a_corner_radius_tr.set_value(inner_radius(a_corner_radius_tr.as<int>(), a_border_thickness.as<int>()));
+	m_titlebar->a_corner_radius_tl.set_value(inner_radius(a_corner_radius_tl.as<double>(), a_border_thickness.as<double>()));
+	m_titlebar->a_corner_radius_tr.set_value(inner_radius(a_corner_radius_tr.as<double>(), a_border_thickness.as<double>()));
 
-	m_start_theme_combobox->a_corner_radius_tl.set_value(7);
-	m_start_theme_combobox->a_corner_radius_tr.set_value(7);
-	m_start_theme_combobox->a_corner_radius_bl.set_value(7);
-	m_start_theme_combobox->a_corner_radius_br.set_value(7);
+	m_start_theme_combobox->a_corner_radius_tl.set_value(7.0);
+	m_start_theme_combobox->a_corner_radius_tr.set_value(7.0);
+	m_start_theme_combobox->a_corner_radius_bl.set_value(7.0);
+	m_start_theme_combobox->a_corner_radius_br.set_value(7.0);
 
-	m_theme_name_line_edit->a_border_thickness.set_value(3);
-	m_theme_name_line_edit->a_corner_radius_tl.set_value(7);
-	m_theme_name_line_edit->a_corner_radius_tr.set_value(7);
-	m_theme_name_line_edit->a_corner_radius_bl.set_value(7);
-	m_theme_name_line_edit->a_corner_radius_br.set_value(7);
-	m_theme_name_line_edit->a_left_padding.set_value(10);
+	m_theme_name_line_edit->a_border_thickness.set_value(3.0);
+	m_theme_name_line_edit->a_corner_radius_tl.set_value(7.0);
+	m_theme_name_line_edit->a_corner_radius_tr.set_value(7.0);
+	m_theme_name_line_edit->a_corner_radius_bl.set_value(7.0);
+	m_theme_name_line_edit->a_corner_radius_br.set_value(7.0);
+	m_theme_name_line_edit->a_left_padding.set_value(10.0);
 }
 
 void CreateNewThemeDialog::init_child_themeable_reference_list()
@@ -204,7 +204,7 @@ bool CreateNewThemeDialog::nativeEvent(const QByteArray& eventType, void* messag
 		}
 
 		*result = 0;
-		const LONG borderWidth = a_border_thickness.as<int>() * devicePixelRatio();;
+		const LONG borderWidth = a_border_thickness.as<double>() * devicePixelRatio();;
 		RECT winrect;
 		GetWindowRect(reinterpret_cast<HWND>(winId()), &winrect);
 
@@ -286,20 +286,20 @@ void CreateNewThemeDialog::paintEvent(QPaintEvent* event)
 
 	bool fill_disabled = a_fill.disabled();
 
-	int border_thickness = a_border_thickness.as<int>();
+	int border_thickness = a_border_thickness.as<double>();
 
-	int margin_left = a_margin_left.as<int>();
-	int margin_top = a_margin_top.as<int>();
-	int margin_right = a_margin_right.as<int>();
-	int margin_bottom = a_margin_bottom.as<int>();
+	int margin_left = a_margin_left.as<double>();
+	int margin_top = a_margin_top.as<double>();
+	int margin_right = a_margin_right.as<double>();
+	int margin_bottom = a_margin_bottom.as<double>();
 
 	int draw_width = width() - margin_left - margin_right;
 	int draw_height = height() - margin_top - margin_bottom;
 
-	int corner_radius_tl = a_corner_radius_tl.as<int>();
-	int corner_radius_tr = a_corner_radius_tr.as<int>();
-	int corner_radius_bl = a_corner_radius_bl.as<int>();
-	int corner_radius_br = a_corner_radius_br.as<int>();
+	int corner_radius_tl = a_corner_radius_tl.as<double>();
+	int corner_radius_tr = a_corner_radius_tr.as<double>();
+	int corner_radius_bl = a_corner_radius_bl.as<double>();
+	int corner_radius_br = a_corner_radius_br.as<double>();
 
 	int tl_background_radius = border_thickness ? inner_radius(corner_radius_tl, border_thickness) : corner_radius_tl;
 	int tr_background_radius = border_thickness ? inner_radius(corner_radius_tr, border_thickness) : corner_radius_tr;
@@ -455,7 +455,7 @@ void CreateNewThemeDialog::setup_layout()
 	inner_layout->setAlignment(m_create_button, Qt::AlignRight);
 
 	// Main Layout
-	int margin = a_border_thickness.as<int>();
+	int margin = a_border_thickness.as<double>();
 
 	m_main_layout->setContentsMargins(margin, margin, margin, margin);
 	m_main_layout->setSpacing(0);
